refactor(discord): bootstrap の memory 記録セットアップを memory-recording.ts へ分離する
